fixed compute and endpoint bugs

+++ nwHacksBackend/tools/setup_tables.js
@@ -20,8 +20,8 @@ conn.connect((err) => {
                             rating float(4) NOT NULL DEFAULT 5,
 							currentRideID int,
 							currentRequestID int,
-                            sumEcoStat float(4) NOT NULL DEFAULT 0,
-                            sumCostSavings float(4) NOT NULL DEFAULT 0,
+                            sumEcoStat double NOT NULL DEFAULT 0,
+                            sumCostSavings double NOT NULL DEFAULT 0,
                             CONSTRAINT pk PRIMARY KEY (id)
                         )`;
 
@@ -38,10 +38,10 @@ conn.connect((err) => {
 							username varchar(255) NOT NULL,
                             status int NOT NULL,
                             created_at datetime NOT NULL DEFAULT '0000-00-00 00:00:00',
-                            startLocLon float(8) NOT NULL,
-                            startLocLat float(8) NOT NULL,
-                            endLocLon float(8) NOT NULL,
-                            endLocLat float(8) NOT NULL,
+                            startLocLon double NOT NULL,
+                            startLocLat double NOT NULL,
+                            endLocLon double NOT NULL,
+                            endLocLat double NOT NULL,
                             expire_at datetime NOT NULL,
                             CONSTRAINT pk PRIMARY KEY (id)
                           )`;
@@ -56,10 +56,10 @@ conn.connect((err) => {
 	const ridesQuery = `CREATE TABLE IF NOT EXISTS rides (
                             id int NOT NULL AUTO_INCREMENT,
                             status int NOT NULL,
-                            startLocLon float(8) NOT NULL,
-                            startLocLat float(8) NOT NULL,
-                            endLocLon float(8) NOT NULL,
-                            endLocLat float(8) NOT NULL,
+                            startLocLon double NOT NULL,
+                            startLocLat double NOT NULL,
+                            endLocLon double NOT NULL,
+                            endLocLat double NOT NULL,
                             user_1 int NOT NULL,
                             user_2 int,
                             user_3 int,
